refactor: Revamp SwapPanel and SwapPanelWrap for improved dialog handling and layout
- Enhanced SwapPanel to conditionally render a dialog for swapping based on screen size, improving user interaction on mobile devices.
- Integrated useIntl for internationalization support in dialog titles and button labels.
- Simplified SwapPanelWrap by removing unnecessary dialog logic and focusing on rendering swap content.
- Updated MobileLayout to streamline content rendering and maintain global visibility of the SwapPanel.
