+++ packages/kit/src/views/Market/MarketDetailV2/components/SwapPanel/SwapPanelWrap.tsx
@@ -1,28 +1,21 @@
-import { useCallback, useEffect, useMemo, useRef } from 'react';
+import { useCallback, useEffect, useMemo } from 'react';
 
 import BigNumber from 'bignumber.js';
-import { useIntl } from 'react-intl';
 
-import { Button, Dialog, View, useMedia } from '@onekeyhq/components';
 import { useActiveAccount } from '@onekeyhq/kit/src/states/jotai/contexts/accountSelector';
-import { EJotaiContextStoreNames } from '@onekeyhq/kit-bg/src/states/jotai/atoms';
-import { ETranslations } from '@onekeyhq/shared/src/locale';
 import { equalTokenNoCaseSensitive } from '@onekeyhq/shared/src/utils/tokenUtils';
 
-import { MarketWatchListProviderMirrorV2 } from '../../../MarketWatchListProviderMirrorV2';
 import { useTokenDetail } from '../../hooks/useTokenDetail';
 
 import { useSpeedSwapActions } from './hooks/useSpeedSwapActions';
 import { useSpeedSwapInit } from './hooks/useSpeedSwapInit';
 import { useSwapPanel } from './hooks/useSwapPanel';
-import { ESwapDirection, type ITradeType } from './hooks/useTradeType';
+import { ESwapDirection } from './hooks/useTradeType';
 import { SwapPanelContent } from './SwapPanelContent';
 
 import type { IToken } from './types';
 
 export function SwapPanelWrap() {
-  const intl = useIntl();
-  const media = useMedia();
   const { activeAccount } = useActiveAccount({ num: 0 });
   const { networkId, tokenDetail } = useTokenDetail();
 
@@ -34,7 +27,6 @@ export function SwapPanelWrap() {
     setPaymentToken,
     paymentToken,
     paymentAmount,
-    setTradeType,
     tradeType,
     setSlippage,
     slippage,
@@ -115,14 +107,6 @@ export function SwapPanelWrap() {
     }
   }, [speedConfig?.slippage, setSlippage]);
 
-  const dialogRef = useRef<ReturnType<typeof Dialog.show>>(undefined);
-
-  useEffect(() => {
-    if (!media.md) {
-      void dialogRef.current?.close();
-    }
-  }, [media.md]);
-
   const handleApprove = useCallback(() => {
     void speedSwapApproveHandler();
   }, [speedSwapApproveHandler]);
@@ -131,7 +115,7 @@ export function SwapPanelWrap() {
     void speedSwapBuildTx();
   }, [speedSwapBuildTx]);
 
-  const swapPanelContent = (
+  return (
     <SwapPanelContent
       priceRate={priceRate}
       swapMevNetConfig={swapMevNetConfig}
@@ -153,36 +137,4 @@ export function SwapPanelWrap() {
       onApprove={handleApprove}
     />
   );
-
-  const showSwapDialog = (tradeTypeValue: ITradeType) => {
-    setTradeType(tradeTypeValue);
-
-    dialogRef.current = Dialog.show({
-      title: intl.formatMessage({ id: ETranslations.global_swap }),
-      renderContent: (
-        <MarketWatchListProviderMirrorV2
-          storeName={EJotaiContextStoreNames.marketWatchListV2}
-        >
-          {swapPanelContent}
-        </MarketWatchListProviderMirrorV2>
-      ),
-      showFooter: false,
-    });
-  };
-
-  if (media.md) {
-    return (
-      <View p="$3">
-        <Button
-          size="large"
-          variant="primary"
-          onPress={() => showSwapDialog(ESwapDirection.BUY)}
-        >
-          {intl.formatMessage({ id: ETranslations.dexmarket_details_trade })}
-        </Button>
-      </View>
-    );
-  }
-
-  return <>{swapPanelContent}</>;
 }
